New projects: one to compile CLI as a DLL and another containing usage example

+++ cli/model.h
@@ -155,8 +155,8 @@ public:
 private:
     bool            m_hasNegativeValues;
     EncodingType    m_encoding;              // io encoding determined based on input file
-    long            m_totalCombinations;     // number of combinations PICT dealt with in this run
-    long            m_remainingCombinations; // number of uncovered combinations (Preview and Approximate)
+    size_t          m_totalCombinations;     // number of combinations PICT dealt with in this run
+    size_t          m_remainingCombinations; // number of uncovered combinations (Preview and Approximate)
 
     void readFile                  ( const std::wstring& filePath );
     bool readModel                 ( const std::wstring& filePath );
